update app before vite

+++ staff/belen-ivars/frontend/app/logic.js
@@ -292,6 +292,7 @@ class Logic {
     deletePost(postId, callback) {
         validateText(postId, 'post id')
 
+
         db.posts.findById(postId, (error, post) => {
             if (error) {
                 callback(error)
@@ -301,19 +302,71 @@ class Logic {
             if (!post) {
 
                 callback(new Error('post not found'))
+
+                return
             }
 
-            db.posts.deleteById(postId, error => {
+            db.users.getAll((error, users) => {
                 if (error) {
                     callback(error)
 
                     return
                 }
 
-                callback(null)
+                const usersWithFav = users.filter((user) => user.favs.includes(postId))
+
+                let count = 0
+
+                if (!usersWithFav.length) {
+                    db.posts.deleteById(postId, error => {
+                        if (error) {
+                            callback(error)
+
+                            return
+                        }
+
+                        callback(null)
+                    })
+
+                    return
+                }
+
+                usersWithFav.forEach(user => {
+
+                    const index = user.favs.indexOf(postId)
+
+                    user.favs.splice(index, 1)
+
+                    db.users.update(user, error => {
+                        if (error) {
+                            callback(error)
+
+                            return
+                        }
+
+                        count++
+
+                        if (count === usersWithFav.length) {
+
+                            db.posts.deleteById(postId, error => { // Abel ho té diferent: (post.id, (error, post) =>)
+                                if (error) {
+                                    callback(error)
+
+                                    return
+                                }
+
+                                callback(null)
+                            })
+                        }
+
+                    })
+                })
+
+
             })
         })
     }
+
 
     retrieveFavPosts(callback) {
         db.users.findById(this.sessionUserId, (error, user) => {
